+++ backend/graph/subscriptions/broadcast.go
@@ -4,13 +4,13 @@ import (
 	gmodel "github.com/semanser/ai-coder/graph/model"
 )
 
-func BroadcastTaskAdded(flowID uint, task *gmodel.Task) {
+func BroadcastTaskAdded(flowID int64, task *gmodel.Task) {
 	if ch, ok := taskAddedSubscriptions[flowID]; ok {
 		ch <- task
 	}
 }
 
-func BroadcastFlowUpdated(flowID uint, flow *gmodel.Flow) {
+func BroadcastFlowUpdated(flowID int64, flow *gmodel.Flow) {
 	if ch, ok := flowUpdatedSubscriptions[flowID]; ok {
 		ch <- flow
 	}
